backend/api/permissions.py
from rest_framework import permissions
from rest_framework.permissions import BasePermission


class AuthorOrReadOnly(BasePermission):
    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True
        if obj.author == request.user:
            return True
