backend/requirements.txt
djangorestframework==3.12.4
Django==2.2.19
python-dotenv==0.20.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
Pillow==9.1.0
djoser==2.1.0
drf_extra_fields==3.4.0
django-filter==21.1